scheduleExecutionTests tests to junit; move some tests to a spec
