+++ rundeckapp/src/test/groovy/rundeck/ScheduledExecutionSpec.groovy
@@ -16,12 +16,16 @@
 
 package rundeck
 
-import spock.lang.Specification
+import grails.test.hibernate.HibernateSpec
+import org.junit.Test
+
+import static org.junit.Assert.*
 
 /**
  * Created by greg on 10/21/15.
  */
-class ScheduledExecutionSpec extends Specification {
+class ScheduledExecutionSpec extends HibernateSpec
+{
     def "has nodes selected by default"() {
         given:
             def se = new ScheduledExecution(nodesSelectedByDefault: value)
@@ -57,4 +61,150 @@ class ScheduledExecutionSpec extends Specification {
             se.options.every { it.scheduledExecution == se }
     }
 
+    void testGenerateJobScheduledName() {
+        when:
+        def ScheduledExecution se = new ScheduledExecution()
+        def props = [jobName: "TestName", project: "TestFrameworkProject", type: "AType", command: "doCommand", argString: "-test", description: "whatever"]
+        se.properties = props
+        se.validate()
+        def StringBuffer sb = new StringBuffer()
+        se.errors.allErrors.each { sb << it.toString() }
+        then:
+        assertTrue "should validate: ${sb.toString()}", se.validate()
+        when:
+        se.save()
+        then:
+        assertEquals 1, ScheduledExecution.count()
+        assertNotNull "id should be set: ${se.id}", se.id
+        assertEquals "incorrect job name: ${se.generateJobScheduledName()}", se.id + ":TestName", se.generateJobScheduledName()
+    }
+
+    void testConstraintsRetry(){
+        when:
+        def ScheduledExecution se = new ScheduledExecution(
+                jobName: "TestName",
+                project: "TestFrameworkProject",
+                argString: "-test",
+                description: "whatever",
+                retry:'123'
+        )
+        assertTrue se.validate()
+        se.retry='${option.retry}'
+        assertTrue se.validate()
+        se.retry='123 '
+        assertFalse se.validate()
+        se.retry='1'
+        assertTrue se.validate()
+        se.retry='0'
+        assertTrue se.validate()
+        se.retry='-2'
+        assertFalse se.validate()
+
+        then:
+        // the above asserts validate the tests
+        1 == 1
+    }
+
+    void testInvalidServerNodeUUID() {
+        when:
+        ScheduledExecution se = createBasicScheduledExecution()
+        se.serverNodeUUID = "blah"
+
+        then:
+        assertFalse(se.validate())
+        assertTrue(se.errors.hasFieldErrors('serverNodeUUID'))
+    }
+
+    private ScheduledExecution createBasicScheduledExecution() {
+        new ScheduledExecution(
+                jobName: "test",
+                groupPath: "",
+                description: "",
+                project: "test",
+                workflow: new Workflow(
+                        commands: [
+                                new CommandExec(adhocRemoteString: "exec")
+                        ]
+                ),
+                options: [],
+        )
+    }
+
+    void testValidateBasic() {
+        when:
+        ScheduledExecution se = createBasicScheduledExecution()
+        then:
+        assertTrue(se.validate())
+    }
+
+    void testDeleteScheduleExecutionWorkflowCascadeAll() {
+
+        when:
+        WorkflowStep workflowStep = new CommandExec([adhocRemoteString: 'test1 buddy', argString: '-delay 12 -monkey cheese -particle'])
+
+        ScheduledExecution se1 = new ScheduledExecution(
+                uuid: 'test1',
+                jobName: 'red color',
+                project: 'Test',
+                groupPath: 'some',
+                description: 'a job',
+                argString: '-a b -c d',
+                workflow: new Workflow(keepgoing: true, commands: [workflowStep]).save(),
+        )
+
+        assert null != se1.save(flush: true)
+
+        then:
+        assertNotNull ScheduledExecution.findById(se1.id)
+        assertNotNull Workflow.findById(se1.workflowId)
+
+        when:
+        se1.delete(flush: true)
+
+        then:
+        assertNull ScheduledExecution.findById(se1.id)
+        assertFalse Workflow.findAll().any {Workflow w -> w.id == se1.workflowId}
+    }
+
+    void testConstraints() {
+        when:
+        def ScheduledExecution se = new ScheduledExecution()
+        def props = [jobName: "TestName", project: "TestFrameworkProject", type: "AType", command: "doCommand", argString: "-test", description: "whatever"]
+        se.properties = props
+        se.validate()
+        def StringBuffer sb = new StringBuffer()
+        se.errors.allErrors.each { sb << it.toString() }
+        assertTrue "ScheduledExecution should validate: ${sb}", se.validate()
+
+        //change values for jobName
+        se.jobName = null
+        assertFalse "ScheduledExecution shouldn't validate", se.validate()
+
+        se.jobName = ""
+        assertFalse "ScheduledExecution shouldn't validate", se.validate()
+
+        List notBlankFields = ['jobName', 'project']
+        notBlankFields.each { key ->
+            se = new ScheduledExecution()
+            se.properties = props
+            assertTrue se.validate()
+            //change values for project
+            se."${key}" = null
+            assertFalse "ScheduledExecution shouldn't validate for null value of ${key}", se.validate()
+            se."${key}" = ""
+            assertFalse "ScheduledExecution shouldn't validate for blank value of ${key}", se.validate()
+        }
+
+        then:
+        // above asserts validate the test
+        1 == 1
+    }
+
+    void testValidateServerNodeUUID() {
+        when:
+        ScheduledExecution se = createBasicScheduledExecution()
+        se.serverNodeUUID = UUID.randomUUID().toString()
+        then:
+        assertTrue(se.validate())
+    }
 }
